Refactoring project, new structure, fix all imports
