+++ app/auth/__init__.py
@@ -1,0 +1,4 @@
+__all__ = "get_current_user,get_current_admin"
+
+
+from .dependencies import get_current_admin, get_current_user
