+++ app/auth/utils.py
@@ -1,4 +1,5 @@
 import os
+from pathlib import Path
 
 from dotenv import load_dotenv
 
@@ -6,12 +7,14 @@ load_dotenv()
 
 
 def get_private_key():
-    with open("jwt-private.pem", "r") as f:
+    file_path = Path(__file__).parent.parent.parent / "jwt-private.pem"
+    with open(file_path, "r") as f:
         return f.read()
 
 
 def get_public_key():
-    with open("jwt-public.pem", "r") as f:
+    file_path = Path(__file__).parent.parent.parent / "jwt-public.pem"
+    with open(file_path, "r") as f:
         return f.read()
