+++ app/auth/views.py
@@ -3,8 +3,8 @@ from typing import Annotated
 from fastapi import APIRouter, Depends, Form, Response, status
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from database.db_helper import db_helper
-from models.user import User
+from app.database.db_helper import db_helper
+from app.models.user import User
 
 from . import service
 from .dependencies import get_current_user
