+++ app/service/views.py
@@ -2,9 +2,9 @@ from fastapi import APIRouter, Depends, File, UploadFile
 from sqlalchemy.ext.asyncio import AsyncSession
 from starlette import status
 
-from auth.dependencies import get_current_admin, get_current_user
-from database.db_helper import db_helper
-from models.user import User
+from app.auth import get_current_admin, get_current_user
+from app.database import db_helper
+from app.models.user import User
 
 from . import service
 from .shema import CreateService, ServiceRespon
