+++ config/README
@@ -1,0 +1,2 @@
+This directory is maintained by the pre-commit project.
+Learn more: https://github.com/pre-commit/pre-commit
